+++ src/App.jsx
@@ -1,5 +1,5 @@
-import TodoList from './TodoList';
-import TodoForm from './TodoForm';
+import TodoList from './features/TodoList/TodoList';
+import TodoForm from './features/TodoForm';
 import './App.css';
 import { useState } from 'react';
